Add hooks for end/beginning of navigableModels
This change will make it easier to add custom behaviour for when we
reach the beginning/end of the current list and click in the `previous`
or `next` buttons.

In most cases, we'll want to go to the next or previous page and keep
the navigation flowing, but it's up to the application using
`ember-cli-paint` to decide what it should do.

Essentially, we are exposing two hooks to handle that,
`getPreviousModelBeforeFirst` and `getNextModelAfterLast`, that are
expected to return a promise that resolves the model that will be used
as the next/previous record.

Here's an example implementation:

```javascript
import Ember from 'ember';
import ModelsNavigationMixin, { Navigator } from 'ember-cli-paint/mixins/models-navigation';

var CustomNavigator = Navigator.extend({
  controller: null,
  projectsController: null,

  model: Ember.computed.oneWay('controller.project'),
  models: Ember.computed.oneWay('projectsController.sortedProjects'),

  getNextModelAfterLast: function() {
    return this.get('projectsController').movePage(1).then((models) => {
      return models.get('firstObject');
    });
  },

  getPreviousModelBeforeFirst: function() {
    return this.get('projectsController').movePage(-1).then((model) => {
      return models.get('lastObject');
    });
  }
});

export default Ember.Controller.extend(ModelsNavigationMixin, {
  project: Ember.computed.oneWay('model.project'),
  projectsController: Ember.inject.controller('projects'),

  navigator: Ember.computed(function() {
    return CustomNavigator.create({
      controller: this,
      projectsController: this.get('projectsController')
    });
  })
}
```

We are also exposing `disablePrevious` and `disableNext` to the
controller. By default these functions will just check if the current
model is the first/last, but we can have custom behaviours per app.